Google Map Problem solved in Lab 4 Stevens Project
This is the file that I changed in order to help students who are currently in China. Since Google Map is not available in China, I use AMap to replace the function of Google Map. Unfortunately, all of the labels in the Map are in Chinese. 

If anybody interested to use this file, you will need to go to the website: lbs.amap.com/dev/id/choose 
and create an account to generate your own key. Then once you obtain the key, you will go into the index.html files to edit and replace "YOUR_KEY" to the actual key you obtain from the website.